add ExampleResult_GetType() function
the reason I change s to success in the previous commit :joy:

Cause `type` looks wierd in vim because it's a golang keyword. Going to avoid that nonsense.

Honestly makes me think if `Type` is the right name for the Result. Maye "Category" or something like that. Something to consider.